pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hn-data-fetcher"
version = "1.1.0"
authors = [
  { name="Adhika Setya Pramudita", email="adhika.setya.p@gmail.com" },
]
description = "A high-performance asynchronous tool for fetching and storing Hacker News items in a SQLite database."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pandas>=1.3.0",
    "sqlalchemy>=1.4.0",
    "pyarrow>=7.0.0",  # Required for parquet support
    "boto3>=1.26.0",  # For S3 operations
    "python-dotenv>=1.0.0",  # For environment variables
    "rich>=13.0.0",  # For beautiful console output
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.0.270",
    "black>=23.3.0",
    "mypy>=1.3.0",
]

[project.urls]
"Homepage" = "https://github.com/adhikasp/hn-data-fetcher"
"Bug Tracker" = "https://github.com/adhikasp/hn-data-fetcher/issues"

[project.scripts]
hn-data-fetcher = "hn_data_fetcher:main"
hn-to-hive = "hn_data_fetcher:hn_to_hive"
upload-to-tigris = "hn_data_fetcher:upload_to_tigris"
delete-from-tigris = "hn_data_fetcher:delete_from_tigris"

[tool.hatch.build.targets.wheel]
packages = ["src/hn_data_fetcher"]

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = [
    "tests",
]

[tool.ruff]
line-length = 88
target-version = "py38" 